refactor(challenge-entity): moves and renames files for consistency and ease of use
