+++ libs/atqr/domain/src/index.ts
@@ -1,4 +1,4 @@
 export * from './lib/player';
-export * from './lib/challenge-entity/challenge-entity';
+export * from './lib/challenge-entity/challenge.entity';
 export * from './lib/EmailService';
 export * from './lib/payment-method';
